requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
ntscraper>=0.3.7
huggingface-hub>=0.20.0
lxml>=5.0.0
